backend/requirements.txt
fastapi>=0.95.0,<0.110.0
uvicorn>=0.22.0,<0.28.0
supabase==0.7.1
httpx>=0.23.0,<0.24.0
pydantic==1.10.8
pydantic[email]==1.10.8
python-multipart>=0.0.5,<0.1.0
python-jose[cryptography]>=3.0.0,<4.0.0
passlib[bcrypt]>=1.7.0,<1.8.0
pytest>=7.0.0,<8.0.0
python-dotenv>=1.0.0,<2.0.0
tenacity>=8.0.0,<9.0.0
psycopg2-binary>=2.9.0,<3.0.0